rename gothamads to intenze
