+++ adapters/intenze/intenzetest/exemplary/banner-app.json
@@ -59,7 +59,7 @@
             "123.123.123.123"
           ]
         },
-        "uri": "http://us-e-node1.gothamads.com/?pass=accountId",
+        "uri": "http://lb-east.intenze.co/?pass=accountId",
         "body": {
           "id": "some-request-id",
           "device": {
@@ -95,7 +95,9 @@
           },
           "tmax": 1000
         },
-        "impIDs":["1"]
+        "impIDs": [
+          "1"
+        ]
       },
       "mockResponse": {
         "status": 200,
@@ -106,7 +108,7 @@
               "bid": [
                 {
                   "id": "a3ae1b4e2fc24a4fb45540082e98e161",
-                  "impid": "test-imp-id",
+                  "impid": "1",
                   "price": 3.5,
                   "adm": "awesome-markup",
                   "adomain": [
@@ -114,11 +116,12 @@
                   ],
                   "crid": "20",
                   "w": 320,
-                  "h": 50
+                  "h": 50,
+                  "mtype": 1
                 }
               ],
-              "mtype": 0,
-              "seat": "gothamads"
+              "type": "banner",
+              "seat": "intenze"
             }
           ],
           "cur": "USD"
@@ -126,10 +129,26 @@
       }
     }
   ],
-  "expectedMakeBidsErrors": [
+  "expectedBidResponses": [
     {
-      "value": "unsupported MType 0",
-      "comparison": "literal"
+      "bids": [
+        {
+          "bid": {
+            "id": "a3ae1b4e2fc24a4fb45540082e98e161",
+            "impid": "1",
+            "price": 3.5,
+            "adm": "awesome-markup",
+            "adomain": [
+              "awesome.com"
+            ],
+            "crid": "20",
+            "w": 320,
+            "h": 50,
+            "mtype": 1
+          },
+          "type": "banner"
+        }
+      ]
     }
   ]
 }
