+++ adapters/intenze/intenzetest/supplemental/invalid-bidder-ext-object.json
