+++ adapters/intenze/intenzetest/supplemental/invalid-intenze-ext-object.json
