+++ adapters/intenze/intenzetest/supplemental/status-code-service-unavailable.json
@@ -35,7 +35,7 @@
   "httpCalls": [
     {
       "expectedRequest": {
-        "uri": "http://us-e-node1.gothamads.com/?pass=accountId",
+        "uri": "http://lb-east.intenze.co/?pass=accountId",
         "body": {
           "id": "some-request-id",
           "imp": [
@@ -64,17 +64,19 @@
           },
           "tmax": 1000
         },
-        "impIDs":["some-impression-id"]
+        "impIDs": [
+          "some-impression-id"
+        ]
       },
       "mockResponse": {
-        "status": 306
+        "status": 503
       }
     }
   ],
   "expectedBidResponses": [],
   "expectedMakeBidsErrors": [
     {
-      "value": "Unexpected status code: 306. Run with request.debug = 1 for more info",
+      "value": "Something went wrong Status Code: [ 503 ] ",
       "comparison": "literal"
     }
   ]
