--- static/bidder-info/gothamads.yaml
@@ -1,14 +1,0 @@
-endpoint: "http://us-e-node1.gothamads.com/?pass={{.AccountID}}"
-maintainer:
-  email: "support@gothamads.com"
-capabilities:
-  app:
-    mediaTypes:
-      - banner
-      - video
-      - native
-  site:
-    mediaTypes:
-      - banner
-      - video
-      - native
